requirements.txt
Flask==2.3.3
pydantic==1.10.13
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
google-generativeai>=0.3.0
